Move common fields to fields
