+++ nap/mapper/models.py
@@ -1,22 +1,20 @@
-import datetime
-
 from django.core.exceptions import ValidationError
 from django.db.models.fields import NOT_PROVIDED
 
 from .base import Mapper, MetaMapper as BaseMetaMapper
-from .fields import field
+from . import fields
 
 
 class Options:
     def __init__(self, meta):
         self.model = getattr(meta, 'model', None)
-        fields = getattr(meta, 'fields', [])
-        if fields != '__all__':
-            fields = list(fields)
-        self.fields = fields
-        self.exclude = list(getattr(meta, 'exclude', []))
+        field_list = getattr(meta, 'fields', [])
+        if field_list != '__all__':
+            field_list = set(field_list)
+        self.fields = field_list
+        self.exclude = set(getattr(meta, 'exclude', []))
         self.required = getattr(meta, 'required', {})
-        self.readonly = getattr(meta, 'readonly', [])
+        self.readonly = set(getattr(meta, 'readonly', []))
 
 
 class MetaMapper(BaseMetaMapper):
@@ -39,7 +37,7 @@ class MetaMapper(BaseMetaMapper):
                 if meta.fields != '__all__' and \
                    f.name not in meta.fields:
                     # Ensure model validation is told to exclude this
-                    meta.exclude.append(f.name)
+                    meta.exclude.add(f.name)
                     continue
 
                 # XXX Magic for field types
@@ -49,7 +47,6 @@ class MetaMapper(BaseMetaMapper):
                     'required': meta.required.get(f.name, not f.blank),
                 }
 
-                field_class = _Field
                 if f.is_relation:
                     kwargs['model'] = f.related_model
                     # OneToOneField or ForeignKey
@@ -59,7 +56,7 @@ class MetaMapper(BaseMetaMapper):
                     elif f.many_to_many and not f.auto_created:
                         field_class = ToManyField
                 else:
-                    field_class = FIELD_MAP.get(f.__class__.__name__, _Field)
+                    field_class = FIELD_MAP.get(f.__class__.__name__, fields.Field)
 
                 attrs[f.name] = field_class(f.name, **kwargs)
             # Inherit
@@ -93,100 +90,7 @@ class ModelMapper(Mapper, metaclass=MetaMapper):
                 self._errors.setdefault(k, []).extend(v)
 
 
-class _Field(field):
-    def __init__(self, attr, *args, **kwargs):
-        self.attr = attr
-        # self.required = kwargs.pop('required', False)
-        # self.default = kwargs.pop('default', NOT_PROVIDED)
-        # self.readonly = kwargs.pop('readonly', False)
-        super().__init__(*args, **kwargs)
-
-    def __get__(self, instance, cls=None):
-        if instance is None:
-            return self
-        value = getattr(instance._obj, self.attr, self.default)
-        if value is NOT_PROVIDED:
-            raise AttributeError("'{}' has no attribute '{}'".format(
-                cls.__name__,
-                self.attr,
-            ))
-        return self.get(value)
-
-    def __set__(self, instance, value):
-        if self.readonly:
-            raise AttributeError('Field {.name} is read-only.'.format(self))
-        value = self.set(value)
-        setattr(instance._obj, self.attr, value)
-
-    def get(self, value):
-        return value
-
-    def set(self, value):
-        return value
-
-
-class BooleanField(_Field):
-    def set(self, value):
-        if value is None:
-            return value
-        if isinstance(value, bool):
-            return value
-        return value.lower() in (1, '1', 't', 'y', 'true')
-
-
-class IntegerField(_Field):
-    def get(self, value):
-        return int(value)
-
-    def set(self, value):
-        return int(value)
-
-
-class FloatField(_Field):
-    def get(self, value):
-        return float(value)
-
-    def set(self, value):
-        return float(value)
-
-
-class TimeField(_Field):
-    def get(self, value):
-        if value is None:
-            return value
-        return value.replace(microsecond=0).isoformat()
-
-    def set(self, value):
-        if value is None or isinstance(value, datetime.time):
-            return value
-            return datetime.datetime.strptime(value, '%H:%M:%S').time()
-
-
-class DateField(_Field):
-    def get(self, value):
-        if value is None:
-            return value
-        return value.isoformat()
-
-    def set(self, value):
-        if value is None or isinstance(value, datetime.date):
-            return value
-        return datetime.datetime.strptime(value, '%Y-%m-%d').date()
-
-
-class DateTimeField(_Field):
-    def get(self, value):
-        if value is None:
-            return value
-        return value.replace(microsecond=0).isoformat(' ')
-
-    def set(self, value):
-        if value is None or isinstance(value, datetime.datetime):
-            return value
-        return datetime.datetime.strptime(value, '%Y-%m-%d %H:%M:%S')
-
-
-class RelatedField(field):
+class RelatedField(fields.field):
     mapper = None
 
     def __init__(self, *args, model=None, mapper=None, **kwargs):
@@ -213,10 +117,10 @@ class ToManyField(RelatedField):
 
 
 FIELD_MAP = {
-    'IntegerField': IntegerField,
-    'FloatField': FloatField,
-    'BooleanField': BooleanField,
-    'DateField': DateField,
-    'TimeField': TimeField,
-    'DateTimeField': DateTimeField,
+    'IntegerField': fields.IntegerField,
+    'FloatField': fields.FloatField,
+    'BooleanField': fields.BooleanField,
+    'DateField': fields.DateField,
+    'TimeField': fields.TimeField,
+    'DateTimeField': fields.DateTimeField,
 }
